Initial implementation of cassandra backups using cassandra-medusa (DM-51254)
Medusa is running as gRPC service in a container on each node. New CLI
is using client module to talk to the gRPC service to make/delete/list
backups. Backups are sent to the S3 bucket, this was tested for bot test
and production clusters. Restore has not been tested yet.